src/app/shared/main-navigation/main-navigation.component.ts
import { Component } from '@angular/core';
import { CommonModule } from '@angular/common';

@Component({
  selector: 'app-main-navigation',
  standalone: true,
  imports: [CommonModule],
  templateUrl: './main-navigation.component.html',
  styleUrls: ['./main-navigation.component.scss'],
})
export class MainNavigationComponent {
  links: string[] = ['Home', 'Services', 'Portfolio', 'Company', 'Blog'];
  isMenuOpen: boolean = false;

  toggleMenu() {
    this.isMenuOpen = !this.isMenuOpen;
  }
}
